# Avoid performance bottleneck corner case in grouping