--- a/src/dataframerow/utils.jl
+++ b/src/dataframerow/utils.jl
@@ -120,7 +120,7 @@ function row_group_slots(cols::Tuple{Vararg{AbstractVector}},
     @assert groups === nothing || length(groups) == length(cols[1])
     rhashes, missings = hashrows(cols, skipmissing)
     # inspired by Dict code from base cf. https://github.com/JuliaData/DataTables.jl/pull/17#discussion_r102481481
-    # but using open addressing with a table with as many slots as rows times 5/4 to avoid performance degradation
+    # but using open addressing with a table with 5/4 as many slots as rows to avoid performance degradation
     # in a corner case of groups having exactly one row
     sz = Base._tablesz((length(rhashes) * 5) >> 2)
     @assert sz >= length(rhashes)